# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nephos API
  version: 0.1.0
paths:
...
  /v1/workspaces/{workspace}/subnets/{id}:
    parameters:
      - $ref: '#/components/parameters/Workspace'
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getSubnet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subnet snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subnet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteSubnet
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Subnet accepted for asynchronous deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subnet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/workspaces/{workspace}/instances:
    parameters:
      - $ref: '#/components/parameters/Workspace'
    post:
      operationId: runInstance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunInstanceRequest'
      responses:
        '201':
          description: Instance accepted for asynchronous creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listInstances
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PageToken'
      responses:
        '200':
          description: Instance snapshots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstancesPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/workspaces/{workspace}/instances/{id}:
    parameters:
      - $ref: '#/components/parameters/Workspace'
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getInstance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Instance snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: terminateInstance
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Instance accepted for asynchronous termination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Workspace:
      name: workspace
      in: path
      required: true
      description: M1 supports only the implicit default workspace
      schema:
        type: string
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Replays identical creates for 24 hours
      schema:
        type: string
        minLength: 1
        maxLength: 255
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    PageToken:
      name: page_token
      in: query
      required: false
      schema:
        type: string
...
  schemas:
...
    Subnet:
      type: object
      additionalProperties: false
      required: [id, name, vpc_id, cidr_block, availability_zone, state, state_reason, generation, observed_generation]
      properties:
        id: {type: string}
        name: {type: string}
        vpc_id: {type: string}
        cidr_block: {type: string}
        availability_zone:
          type: string
          enum: [local-1a, local-1b, local-1c]
        state:
          type: string
          enum: [pending, available, deleting, failed]
        state_reason: {type: string}
        generation: {type: integer, format: int64}
        observed_generation: {type: integer, format: int64}
...
    RunInstanceRequest:
      type: object
      additionalProperties: false
      required: [name, subnet_id]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        subnet_id: {type: string}
    Instance:
      type: object
      additionalProperties: false
      required: [id, name, subnet_id, private_ip, eni_id, instance_type, state, state_reason, generation, observed_generation]
      properties:
        id: {type: string}
        name: {type: string}
        subnet_id: {type: string}
        private_ip: {type: string}
        eni_id: {type: string}
        instance_type:
          type: string
          enum: [t3.micro]
        state:
          type: string
          enum: [pending, running, stopping, stopped, shutting-down, terminated, failed]
        state_reason: {type: string}
        generation: {type: integer, format: int64}
        observed_generation: {type: integer, format: int64}
    InstancesPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Instance'
        next_page_token: {type: string}
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        resource_id: {type: string}
        docs_url: {type: string}
  responses:
    BadRequest:
      description: BadRequest error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: NotFound error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'