# internal/store/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"database/sql"
)

type Eni struct {
	ID                 string
	WorkspaceID        string
	InstanceID         string
	SubnetID           string
	ShortIndex         int64
	PrivateIp          string
	MacAddress         string
	Generation         int64
	ObservedGeneration int64
	State              string
	StateReason        string
	CreatedAt          int64
	UpdatedAt          int64
}

type Event struct {
	ID           int64
	WorkspaceID  string
	ResourceType string
	ResourceID   string
	Action       string
	State        string
	Generation   int64
	Message      string
	CreatedAt    int64
}

type IdempotencyRequest struct {
	WorkspaceID  string
	Operation    string
	Key          string
	PayloadHash  string
	ResourceID   string
	CreatedAt    int64
	ExpiresAt    int64
	ResponseJson string
}

type Instance struct {
	ID                 string
	WorkspaceID        string
	SubnetID           string
	ShortIndex         int64
	Name               string
	Generation         int64
	ObservedGeneration int64
	State              string
	StateReason        string
	CreatedAt          int64
	UpdatedAt          int64
	DeletionRequested  int64
	RuntimeID          sql.NullString
}

type KernelIndex struct {
	ID           int64
	ResourceKind string
	ResourceID   string
	AllocatedAt  int64
}

type SchemaMigration struct {
	Version   int64
	AppliedAt int64
}

type Subnet struct {
	ID                 string
	WorkspaceID        string
	VpcID              string
	ShortIndex         int64
	Name               string
	CidrBlock          string
	AvailabilityZone   string
	Generation         int64
	ObservedGeneration int64
	State              string
	StateReason        string
	CreatedAt          int64
	UpdatedAt          int64
	DeletionRequested  int64
}

type Vpc struct {
	ID                 string
	WorkspaceID        string
	ShortIndex         int64
	Name               string
	CidrBlock          string
	Generation         int64
	ObservedGeneration int64
	State              string
	StateReason        string
	CreatedAt          int64
	UpdatedAt          int64
	DeletionRequested  int64
}

type Workspace struct {
	ID        string
	CreatedAt int64
}